sequence_labeling/requirement.txt
absl-py==0.8.1
astor==0.8.0
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.0
protobuf==3.10.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
configparser==4.0.2
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.0
pandas==0.24.0
Pillow==6.2.1
protobuf==3.10.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
urllib3==1.25.7
uWSGI==2.0.18
Werkzeug==0.16.0
wrapt==1.11.2
